+++ ui/app/components/app/asset-list-item/asset-list-item.js
@@ -2,9 +2,9 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import classnames from 'classnames'
 import Identicon from '../../ui/identicon'
+import ListItem from '../../ui/list-item'
 
 const AssetListItem = ({
-  children,
   className,
   'data-testid': dataTestId,
   iconClassName,
@@ -12,32 +12,31 @@ const AssetListItem = ({
   tokenAddress,
   tokenImage,
   warning,
+  primary,
+  secondary,
 }) => {
   return (
-    <div
-      className={classnames('asset-list-item__container', className)}
+    <ListItem
+      className={classnames('asset-list-item', className)}
       data-testid={dataTestId}
+      title={primary}
+      titleIcon={warning}
+      subtitle={secondary}
       onClick={onClick}
-    >
-      <Identicon
-        className={iconClassName}
-        diameter={32}
-        address={tokenAddress}
-        image={tokenImage}
-      />
-      <div
-        className="asset-list-item__balance"
-      >
-        { children }
-      </div>
-      { warning }
-      <i className="fas fa-chevron-right asset-list-item__chevron-right" />
-    </div>
+      icon={(
+        <Identicon
+          className={iconClassName}
+          diameter={32}
+          address={tokenAddress}
+          image={tokenImage}
+        />
+      )}
+      rightContent={<i className="fas fa-chevron-right asset-list-item__chevron-right" />}
+    />
   )
 }
 
 AssetListItem.propTypes = {
-  children: PropTypes.node.isRequired,
   className: PropTypes.string,
   'data-testid': PropTypes.string,
   iconClassName: PropTypes.string,
@@ -45,6 +44,8 @@ AssetListItem.propTypes = {
   tokenAddress: PropTypes.string,
   tokenImage: PropTypes.string,
   warning: PropTypes.node,
+  primary: PropTypes.string,
+  secondary: PropTypes.string,
 }
 
 AssetListItem.defaultProps = {
