--- ui/app/components/app/asset-list/asset-list.scss
@@ -1,13 +1,0 @@
-.asset-list {
-  &__primary-amount {
-    color: $Black-100;
-    font-size: 16px;
-    height: 16px;
-  }
-
-  &__secondary-amount {
-    color: $Grey-500;
-    margin-top: 6px;
-    font-size: 14px;
-  }
-}
