+++ ui/app/components/app/index.scss
@@ -6,8 +6,6 @@
 
 @import 'app-header/index';
 
-@import 'asset-list/asset-list';
-
 @import 'asset-list-item/asset-list-item';
 
 @import '../ui/breadcrumbs/index';
