+++ ui/app/components/app/token-cell/token-cell.scss
@@ -1,95 +1,5 @@
-$wallet-balance-breakpoint: 890px;
-$wallet-balance-breakpoint-range: "screen and (min-width: #{$break-large}) and (max-width: #{$wallet-balance-breakpoint})";
-
 .token-cell {
-  position: relative;
-
-  &__token-balance {
-    margin-right: 4px;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    min-width: 0;
-    max-width: 100%;
-  }
-
-  &__token-balance, &__token-symbol {
-    font-size: 16px;
-    flex: 0 0 auto;
-    color: $Black-100;
-  }
-
-  &__fiat-amount {
-    margin-top: 6px;
-    font-size: 14px;
-    width: 100%;
-    text-transform: uppercase;
+  &--outdated .list-item__heading {
     color: $Grey-500;
-  }
-
-  &--outdated &__icon {
-    opacity: 0.5
-  }
-  &--outdated &__balance-wrapper {
-    opacity: 0.5
-  }
-
-  &__balance-wrapper {
-    flex: 1;
-    flex-flow: row wrap;
-    display: flex;
-    min-width: 0;
-  }
-
-  &__outdated-icon {
-    color: $warning-yellow;
-    display: block;
-    padding: 0 10px;
-  }
-
-  &__outdated-tooltip {
-    width: 260px;
-  }
-}
-
-.token-menu-dropdown {
-  width: 80%;
-  position: absolute;
-  top: 52px;
-  right: 25px;
-  z-index: 2000;
-
-  @media #{$wallet-balance-breakpoint-range} {
-    right: 18px;
-  }
-
-  &__close-area {
-    position: fixed;
-    top: 0;
-    left: 0;
-    z-index: 2100;
-    width: 100%;
-    height: 100%;
-    cursor: default;
-  }
-
-  &__container {
-    padding: 16px;
-    z-index: 2200;
-    position: relative;
-  }
-
-  &__options {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-  }
-
-  &__option {
-    color: $white;
-    font-family: Roboto;
-    font-size: 16px;
-    line-height: 21px;
-    text-align: center;
   }
 }
