+++ ui/app/components/ui/list-item/index.scss
@@ -18,7 +18,7 @@
   }
 
   &__col {
-    align-self: flex-start;
+    align-self: center;
     &-main {
       flex-grow: 1;
     }
@@ -32,7 +32,6 @@
     &-wrap {
       display: inline-block;
       position: absolute;
-      top: 2px;
       width: 16px;
       height: 16px;
       margin-left: 8px;
