+++ ui/app/components/ui/list-item/list-item.component.js
@@ -2,11 +2,22 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import classnames from 'classnames'
 
-export default function ListItem ({ title, subtitle, onClick, subtitleStatus, children, titleIcon, icon, rightContent, className }) {
+export default function ListItem ({
+  title,
+  subtitle,
+  onClick,
+  subtitleStatus,
+  children,
+  titleIcon,
+  icon,
+  rightContent,
+  className,
+  'data-testid': dataTestId,
+}) {
   const primaryClassName = classnames('list-item', className)
 
   return (
-    <div className={primaryClassName} onClick={onClick}>
+    <div className={primaryClassName} onClick={onClick} data-testid={dataTestId}>
       {icon && (
         <div className="list-item__col list-item__icon">
           {icon}
@@ -48,4 +59,5 @@ ListItem.propTypes = {
   rightContent: PropTypes.node,
   className: PropTypes.string,
   onClick: PropTypes.func,
+  'data-testid': PropTypes.string,
 }
